pyproject.toml
[tool.poetry]
name = "gmailhelper"
version = "0.1.0"
description = "A simple python project to manage gmail programmatically"
authors = ["badri <badri.twok@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
black = "^25.1.0"
isort = "^6.0.1"
pydantic = "^2.11.7"
google-api-python-client = "^2.178.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.2"
fastapi = "^0.116.1"
dependency-injector = "^4.48.1"
httpx = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
